tests/pyproject.toml
[project]
name = "tests"
dynamic = ["version"]
requires-python = ">=3.13"
dependencies = ["pytest==8.3.4", "pytest-playwright==0.6.2", "requests==2.32.3", "pytest-rerunfailures==15.0.0"]

[project.optional-dependencies]
dev = ["zizmor==1.3.0", "ruff==0.9.4"]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]

ignore = [
  "COM812",  # Ignore due to conflict with Ruff formatter
  "ISC001",  # Ignore due to conflict with Ruff formatter
  "PLR2004", # Ignore magic value
  "D104",    # Ignore missing docstring in public package
  "S101",    # Ignore use of assert
]

fixable = ["ALL"]
unfixable = []

exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
]

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google"
